BigDecimal problem solved. add classes in client namespace

+++ src/Core/Model/Clients/Base/AbstractToken.cs
@@ -1,5 +1,6 @@
 ﻿using Org.BouncyCastle.Math;
 using Org.BouncyCastle.Utilities;
+using ThorClient.Numerics;
 
 namespace ThorClient.Core.Model.Clients.Base
 {
@@ -8,8 +9,8 @@ namespace ThorClient.Core.Model.Clients.Base
         public static AbstractToken VET { get; } = new AbstractToken("VET", 18);
 
         public string Name { get; set; }
-        public int Precision { get; set; }
-        public int Scale { get; set; }
+        public BigDecimal Precision { get; set; }
+        public BigDecimal Scale { get; set; }
 
         protected AbstractToken(string name, int unit)
         {
